% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcpp11unwind-package.R
\docType{package}
\name{testcpp11unwind-package}
\alias{testcpp11unwind}
\alias{testcpp11unwind-package}
\title{testcpp11unwind: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/testcpp11unwind}
  \item Report bugs at \url{https://github.com/DavisVaughan/testcpp11unwind/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
